requirements.txt
crowdstrike-falconpy
python-dotenv
rich
